refactor(stringcalculator): 문자열 null 및 빈 문자열 체크 기능을 StringHelper Class로 분리
