+++ src/main/java/stringcalculator/StringHelper.java
@@ -1,0 +1,12 @@
+package stringcalculator;
+
+public class StringHelper {
+
+    private StringHelper() {
+        throw new IllegalStateException("Utility class");
+    }
+
+    public static boolean isNullOrEmpty(String input) {
+        return input == null || input.isEmpty();
+    }
+}
